Move convertBabelLoc
