+++ packages/shared/scope-hoisting/package.json
@@ -25,6 +25,7 @@
     "@babel/template": "^7.2.2",
     "@babel/traverse": "^7.2.3",
     "@babel/types": "^7.3.3",
+    "@parcel/babel-ast-utils": "2.0.0-beta.1",
     "@parcel/babylon-walk": "2.0.0-beta.1",
     "@parcel/diagnostic": "2.0.0-beta.1",
     "@parcel/source-map": "2.0.0-alpha.4.13",
