+++ packages/shared/scope-hoisting/src/utils.js
@@ -290,13 +290,7 @@ export function getThrowableDiagnosticForNode(
   if (loc) {
     diagnostic.codeFrame = {
       codeHighlights: {
-        start: {
-          line: loc.start.line,
-          column: loc.start.column + 1,
-        },
-        // - Babel's columns are exclusive, ours are inclusive (column - 1)
-        // - Babel has 0-based columns, ours are 1-based (column + 1)
-        // = +-0
+        start: loc.start,
         end: loc.end,
       },
     };
@@ -304,21 +298,4 @@ export function getThrowableDiagnosticForNode(
   return new ThrowableDiagnostic({
     diagnostic,
   });
-}
-
-export function convertBabelLoc(loc: ?BabelSourceLocation): ?SourceLocation {
-  if (!loc || !loc.filename) return null;
-
-  let {filename, start, end} = loc;
-  return {
-    filePath: path.normalize(filename),
-    start: {
-      line: start.line,
-      column: start.column,
-    },
-    end: {
-      line: end.line,
-      column: end.column,
-    },
-  };
 }
